// src/components/common/types.ts
/**
 * Common.styles
 * START
 */

export interface ImageContainerProps {
  width?: number;
  height?: number;
}

export interface ImageProps {
  borderRadius?: number;
}

export interface ButtonProps {
  width?: number;
  height?: number;
  background?: string;
  color?: string;
  borderColor?: string;
  borderRadius?: number;
}

export interface IconButtonProps {
  width?: number;
  height?: number;
}

/**
 * END
 */